test-requirements.txt
# base requirements
-r requirements.txt

# test requirements
coveralls
django-debug-toolbar
flake8